QueryStatusActionHandler: extract from DetoxActionHandlers.
