+++ detox/android/detox/src/full/java/com/wix/detox/adapters/server/DetoxActionHandlers.kt
@@ -2,12 +2,10 @@ package com.wix.detox.adapters.server
 
 import android.content.Context
 import android.util.Log
-import androidx.test.espresso.IdlingResource
 import com.wix.detox.TestEngineFacade
 import com.wix.detox.common.extractRootCause
 import com.wix.detox.instruments.DetoxInstrumentsException
 import com.wix.detox.instruments.DetoxInstrumentsManager
-import com.wix.detox.reactnative.idlingresources.DescriptiveIdlingResource
 import com.wix.invoke.MethodInvocation
 import org.json.JSONObject
 import java.lang.reflect.InvocationTargetException
@@ -91,101 +89,6 @@ class CleanupActionHandler(
             testEngineFacade.resetReactNative()
         }
         outboundServerAdapter.sendMessage("cleanupDone", emptyMap(), messageId)
-    }
-}
-
-class QueryStatusActionHandler(
-        private val outboundServerAdapter: OutboundServerAdapter,
-        private val testEngineFacade: TestEngineFacade)
-    : DetoxActionHandler {
-
-    override fun handle(params: String, messageId: Long) {
-        val data = mutableMapOf<String, Any?>()
-        data["status"] = formatStatus(testEngineFacade.getBusyIdlingResources())
-
-        outboundServerAdapter.sendMessage("currentStatusResult", data, messageId)
-    }
-
-    private fun formatStatus(busyResources: List<IdlingResource>): Map<String, Any> {
-        if (busyResources.isEmpty()) {
-            return mapOf("app_status" to "idle")
-        }
-
-        val status = mutableMapOf<String, Any>()
-        status["app_status"] = "busy"
-        status["busy_resources"] = busyResources.map{ formatResource(it) }
-
-        return status
-    }
-
-    private fun formatResource(resource: IdlingResource): Map<String, Any> {
-        if (resource is DescriptiveIdlingResource) {
-            return resource.getDescription().json()
-        }
-
-        if (resource.javaClass.name.contains("LooperIdlingResource")) {
-            return formatLooperResourceFromName(resource.name)
-        }
-
-        return mapOf<String, Any>(
-                "name" to "unknown",
-                "description" to mapOf<String, Any>(
-                        "identifier" to resource.name
-                )
-        )
-    }
-
-    private fun formatLooperResourceFromName(resourceName: String): Map<String, Any> {
-        return when {
-            isJSCodeExecution(resourceName) -> {
-                formatLooperResource(
-                    "\"${resourceName}\" (JS Thread)",
-                    "JavaScript code"
-                )
-            }
-            isNativeCodeExecution(resourceName) -> {
-                formatLooperResource(
-                    "\"${resourceName}\" (Native Modules Thread)",
-                    "native module calls"
-                )
-            }
-            else -> {
-                formatLooperResource(
-                    "\"${resourceName}\""
-                )
-            }
-        }
-    }
-
-    /**
-     * @see URL https://reactnative.dev/docs/profiling
-     */
-    private fun isJSCodeExecution(looperName: String): Boolean {
-        return looperName.contains("mqt_js")
-    }
-
-    private fun formatLooperResource(thread: String, executionType: String? = null): Map<String, Any> {
-        return mapOf<String, Any>(
-                "name" to "looper",
-                "description" to
-                        if (executionType != null) {
-                            mapOf<String, Any>(
-                                    "thread" to thread,
-                                    "execution_type" to executionType
-                            )
-                        } else {
-                            mapOf<String, Any>(
-                                    "thread" to thread
-                            )
-                        }
-        )
-    }
-
-    /**
-     * @see URL https://reactnative.dev/docs/profiling
-     */
-    private fun isNativeCodeExecution(looperName: String): Boolean {
-        return looperName.contains("mqt_native")
     }
 }
