+++ detox/android/detox/src/testFull/java/com/wix/detox/adapters/server/DetoxActionHandlersSpec.kt
@@ -1,14 +1,11 @@
 package com.wix.detox.adapters.server
 
 import android.content.Context
-import androidx.test.espresso.IdlingResource
 import com.nhaarman.mockitokotlin2.*
 import com.wix.detox.TestEngineFacade
 import com.wix.detox.UTHelpers.yieldToOtherThreads
 import com.wix.detox.instruments.DetoxInstrumentsException
 import com.wix.detox.instruments.DetoxInstrumentsManager
-import com.wix.detox.reactnative.idlingresources.DescriptiveIdlingResource
-import com.wix.detox.reactnative.idlingresources.IdlingResourceDescription
 import com.wix.invoke.MethodInvocation
 import org.assertj.core.api.Assertions
 import org.json.JSONObject
@@ -94,94 +91,6 @@ object DetoxActionHandlersSpec : Spek({
                 yieldToOtherThreads(executor)
                 verify(testEngineFacade, times(1)).reloadReactNative(eq(appContext))
                 verify(outboundServerAdapter, times(1)).sendMessage(any(), any(), any())
-            }
-        }
-
-        describe("query status") {
-            fun queryStatusHandler() = QueryStatusActionHandler(outboundServerAdapter, testEngineFacade)
-
-            it("should return idle app status") {
-                queryStatusHandler().handle(params, messageId)
-                val expectedData =  mapOf("status" to mapOf("app_status" to "idle"))
-                verify(outboundServerAdapter).sendMessage(eq("currentStatusResult"), eq(expectedData), eq(messageId))
-            }
-
-            fun createMockedDescriptiveResource(description: IdlingResourceDescription):
-                    DescriptiveIdlingResource {
-                val resource: DescriptiveIdlingResource = mock()
-                whenever(resource.getDescription()).thenReturn(description)
-                return resource
-            }
-
-            it("should return busy app status with descriptive resource") {
-                val fakeDescription = IdlingResourceDescription.Builder()
-                    .name("foo")
-                    .addDescription("bar", "baz")
-                    .addDescription("qux", "quux")
-                    .build();
-
-                val busyResources: List<IdlingResource> = listOf(
-                    createMockedDescriptiveResource(fakeDescription)
-                )
-                whenever(testEngineFacade.getBusyIdlingResources()).thenReturn(busyResources)
-
-                queryStatusHandler().handle(params, messageId)
-
-                val expectedBusyResourceDescription = listOf(
-                    mapOf("name" to "foo", "description" to mapOf("bar" to "baz", "qux" to "quux"))
-                )
-                val expectedData =  mapOf("status" to mapOf("app_status" to "busy", "busy_resources" to expectedBusyResourceDescription))
-                verify(outboundServerAdapter).sendMessage(eq("currentStatusResult"), eq(expectedData), eq(messageId))
-            }
-
-            abstract class LooperIdlingResource: IdlingResource {
-                abstract val thread: String?
-            }
-
-            fun createMockedLooperIdlingResource(resourceName: String): LooperIdlingResource {
-                val resource: LooperIdlingResource = mock()
-                whenever(resource.name).thenReturn(resourceName)
-                return resource
-            }
-
-            it("should return busy app status with looper resources") {
-                val busyResources: List<IdlingResource> = listOf(
-                    createMockedLooperIdlingResource("mqt_js"),
-                    createMockedLooperIdlingResource("mqt_native"),
-                    createMockedLooperIdlingResource("unmapped")
-                )
-                whenever(testEngineFacade.getBusyIdlingResources()).thenReturn(busyResources)
-
-                queryStatusHandler().handle(params, messageId)
-
-                val expectedBusyResourceDescription = listOf(
-                    mapOf("name" to "looper", "description" to mapOf("thread" to "\"mqt_js\" (JS Thread)", "execution_type" to "JavaScript code")),
-                    mapOf("name" to "looper", "description" to mapOf("thread" to "\"mqt_native\" (Native Modules Thread)", "execution_type" to "native module calls")),
-                    mapOf("name" to "looper", "description" to mapOf("thread" to "\"unmapped\""))
-                )
-                val expectedData =  mapOf("status" to mapOf("app_status" to "busy", "busy_resources" to expectedBusyResourceDescription))
-                verify(outboundServerAdapter).sendMessage(eq("currentStatusResult"), eq(expectedData), eq(messageId))
-            }
-
-            fun createMockedIdlingResource(name: String): IdlingResource {
-                val resource: IdlingResource = mock()
-                whenever(resource.name).thenReturn(name)
-                return resource
-            }
-
-            it("should return busy app status with unknown resource") {
-                val busyResources: List<IdlingResource> = listOf(
-                    createMockedIdlingResource( "quux")
-                )
-                whenever(testEngineFacade.getBusyIdlingResources()).thenReturn(busyResources)
-
-                queryStatusHandler().handle(params, messageId)
-
-                val expectedBusyResourceDescription = listOf(
-                    mapOf("name" to "unknown", "description" to mapOf("identifier" to "quux")),
-                )
-                val expectedData =  mapOf("status" to mapOf("app_status" to "busy", "busy_resources" to expectedBusyResourceDescription))
-                verify(outboundServerAdapter).sendMessage(eq("currentStatusResult"), eq(expectedData), eq(messageId))
             }
         }
